refactor(useCurrentFolderId): Extract from selectors into hooks
