+++ src/drive/hooks/index.js
@@ -1,1 +1,2 @@
 export { default as useCurrentFileId } from './useCurrentFileId'
+export { default as useCurrentFolderId } from './useCurrentFolderId'
