+++ src/drive/hooks/useCurrentFolderId.jsx
@@ -1,0 +1,18 @@
+import { ROOT_DIR_ID, TRASH_DIR_ID } from 'drive/constants/config'
+import { useParams, useLocation } from 'react-router-dom'
+
+const useCurrentFolderId = () => {
+  const { folderId } = useParams()
+  const { pathname } = useLocation()
+
+  if (folderId) {
+    return folderId
+  } else if (pathname === '/folder') {
+    return ROOT_DIR_ID
+  } else if (pathname === '/trash') {
+    return TRASH_DIR_ID
+  }
+  return null
+}
+
+export default useCurrentFolderId
