+++ src/drive/web/modules/navigation/duck/actions.jsx
@@ -7,10 +7,8 @@ import Alerter from 'cozy-ui/transpiled/react/Alerter'
 import { addToUploadQueue } from 'drive/web/modules/upload'
 import logger from 'lib/logger'
 import QuotaAlert from 'drive/web/modules/upload/QuotaAlert'
-import {
-  useCurrentFolderId,
-  getFolderContent
-} from 'drive/web/modules/selectors'
+import { getFolderContent } from 'drive/web/modules/selectors'
+import { useCurrentFolderId } from 'drive/hooks'
 import { createEncryptedDir } from 'drive/lib/encryption'
 import { logException } from 'drive/lib/reporter'
