+++ src/drive/web/modules/views/Public/index.jsx
@@ -37,10 +37,10 @@ import FolderViewBreadcrumb from '../Folder/FolderViewBreadcrumb'
 import PublicToolbar from 'drive/web/modules/public/PublicToolbar'
 import PublicViewer from 'drive/web/modules/viewer/PublicViewer'
 import {
-  useCurrentFolderId,
   useDisplayedFolder,
   useParentFolder
 } from 'drive/web/modules/selectors'
+import { useCurrentFolderId } from 'drive/hooks'
 import { useExtraColumns } from 'drive/web/modules/certifications/useExtraColumns'
 import { makeExtraColumnsNamesFromMedia } from 'drive/web/modules/certifications'
 import FabWithMenuContext from 'drive/web/modules/drive/FabWithMenuContext'
